Solved MongoBinData Deprecated Exception
Solved an issue caused by and Exception raised to a Deprecated use of the MongoBinData constructor.
The error message is:
Deprecated: MongoBinData::__construct() [<a href='mongobindata.--construct'>mongobindata.--construct</a>]: The default value for type will change to 0 in the future. Please pass in '2' explicitly. in [...]/vendor/bundles/Sonata/CacheBundle/Adapter/MongoCache.php line 93